split filesystem and non fs tests
